Modify to Double instead of Float and adapt doc

+++ src/main/java/com/danielpsf/labs/field/statistic/domain/Statistic.java
@@ -18,5 +18,5 @@ public class Statistic {
 
     @Temporal(TemporalType.TIMESTAMP)
     private Date occurrenceAt;
-    private Float vegetation;
+    private Double vegetation;
 }
